American Leaf Spot

Compute the percentage of Coffea leaves dying from American Leaf Spot disease, drought excluded.

ALS(Elevation, SlopeAzimut = 0, Slope = 0, RowDistance = 1.5,
  Shade = 0, CanopyHeight.Coffee = 2, Fertilization = 3,
  ShadeType = 5, CoffeePruning = c("tree", "row", "block", NULL),
  df_rain)

Arguments

Elevation
Site elevation (m.a.s.l)

SlopeAzimut
Slope azimuth (degree)

Slope
Slope percentage (\(\%\))

RowDistance
Coffee rows distance (m)

Shade
Shade percentage (\(\%\))

CanopyHeight.Coffee
Coffee Height (m)

Fertilization
N fertilization per year

ShadeType
Shade type:

1

Legume only

2

Bananas and legume

3

Bananas and other plants

4

Fruit and forest trees only

5

No shade (Full sun)

CoffeePruning
Character specifying the pruning management. Values: tree, row, block or NULL.

df_rain
Data frame with DOY, year and Rain (mm) values

Value

ALS

Percentage of dead leaves by ALS by day (\(\% day^{-1}\))

Note

It is good practice to use shade tree transmittance to compute Shade percentage (\(Shade= 1-Transmittance\)).

References

Avelino et al. (2007) Topography and Crop Management Are Key Factors for the Development of American Leaf Spot Epidemics on Coffee in Costa Rica File: "Ia_digitized from Avelino 2007 - JA_lineaire.xlsx"

Examples

# df_rain has this structure :
df_rain= data.frame(DOY= 1:365, year= rep(2018,365), Rain= rnorm(n = 365,mean = 5, sd = 1))
ALS(Elevation = 1000, df_rain= df_rain)

#>   [1] 0.000000e+00 2.145489e-03 2.184525e-03 2.224272e-03 2.264742e-03
#>   [6] 2.305948e-03 2.347904e-03 2.390623e-03 2.434120e-03 2.478408e-03
#>  [11] 2.523501e-03 2.569416e-03 2.616165e-03 2.663765e-03 2.712232e-03
#>  [16] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [21] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [26] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [31] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [41] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [51] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [56] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [61] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [66] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [71] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [76] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [81] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [86] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [91] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [101] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [106] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [111] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [116] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [121] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [126] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [131] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [136] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [141] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [146] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [151] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [156] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [161] 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#> [166] 3.202000e-03 5.825924e-05 5.931925e-05 6.039854e-05 6.149747e-05
#> [171] 6.261640e-05 6.375568e-05 6.491569e-05 6.609681e-05 6.729942e-05
#> [176] 6.852390e-05 6.977067e-05 7.104012e-05 7.233267e-05 7.364874e-05
#> [181] 7.498875e-05 7.635315e-05 7.774237e-05 7.915686e-05 8.059709e-05
#> [186] 8.206353e-05 8.355664e-05 8.507693e-05 8.662487e-05 8.820098e-05
#> [191] 8.980576e-05 9.143975e-05 9.310346e-05 9.479744e-05 9.652225e-05
#> [196] 9.827844e-05 1.000666e-04 1.018873e-04 1.037411e-04 1.056286e-04
#> [201] 1.075505e-04 1.095073e-04 1.114998e-04 1.135284e-04 1.155941e-04
#> [206] 1.176973e-04 1.198387e-04 1.220191e-04 1.242392e-04 1.264997e-04
#> [211] 1.288013e-04 1.311448e-04 1.335310e-04 1.359605e-04 1.384343e-04
#> [216] 1.409530e-04 1.435176e-04 1.461289e-04 1.487876e-04 1.514948e-04
#> [221] 1.542512e-04 1.570577e-04 1.599153e-04 1.628249e-04 1.657875e-04
#> [226] 1.688039e-04 1.718752e-04 1.750024e-04 1.781865e-04 1.814286e-04
#> [231] 1.847296e-04 1.880907e-04 1.915129e-04 1.949974e-04 1.985454e-04
#> [236] 2.021578e-04 2.058360e-04 2.095811e-04 2.133944e-04 2.172770e-04
#> [241] 2.212303e-04 2.252555e-04 2.293539e-04 2.335269e-04 2.377759e-04
#> [246] 2.421021e-04 2.465071e-04 2.509922e-04 2.555589e-04 2.602087e-04
#> [251] 2.649431e-04 2.697637e-04 2.746719e-04 2.796695e-04 2.847580e-04
#> [256] 2.899390e-04 2.952144e-04 3.005857e-04 3.060547e-04 3.116233e-04
#> [261] 3.172932e-04 3.230662e-04 3.289443e-04 3.349293e-04 3.410232e-04
#> [266] 3.472280e-04 3.535457e-04 3.599783e-04 3.665280e-04 3.731969e-04
#> [271] 3.799870e-04 3.869008e-04 3.939403e-04 4.011079e-04 4.084059e-04
#> [276] 4.158367e-04 4.234027e-04 4.311064e-04 4.389502e-04 4.469367e-04
#> [281] 4.550686e-04 4.633484e-04 4.717789e-04 4.803627e-04 4.891027e-04
#> [286] 4.980018e-04 5.070627e-04 5.162886e-04 5.256823e-04 5.352469e-04
#> [291] 5.449855e-04 5.549013e-04 5.649975e-04 5.752775e-04 5.857444e-04
#> [296] 5.964018e-04 6.072532e-04 6.183019e-04 6.295517e-04 6.410062e-04
#> [301] 6.526690e-04 6.645441e-04 6.766353e-04 6.889464e-04 7.014815e-04
#> [306] 7.142447e-04 7.272402e-04 7.404720e-04 7.539446e-04 7.676624e-04
#> [311] 7.816297e-04 7.958512e-04 8.103314e-04 8.250751e-04 8.400871e-04
#> [316] 8.553722e-04 8.709354e-04 8.867817e-04 9.029164e-04 9.193446e-04
#> [321] 9.360718e-04 9.531033e-04 9.704446e-04 9.881015e-04 1.006080e-03
#> [326] 1.024385e-03 1.043023e-03 1.062001e-03 1.081323e-03 1.100998e-03
#> [331] 1.121030e-03 1.141427e-03 1.162195e-03 1.183340e-03 1.204871e-03
#> [336] 1.226793e-03 1.249114e-03 1.271841e-03 1.294982e-03 1.318544e-03
#> [341] 1.342534e-03 1.366961e-03 1.391832e-03 1.417156e-03 1.442941e-03
#> [346] 1.469195e-03 1.495926e-03 1.523144e-03 1.550857e-03 1.579074e-03
#> [351] 1.607805e-03 1.637058e-03 1.666844e-03 1.697172e-03 1.728051e-03
#> [356] 1.759492e-03 1.791506e-03 1.824102e-03 1.857290e-03 1.891083e-03
#> [361] 1.925491e-03 1.960524e-03 1.996195e-03 2.032515e-03 2.069496e-03