Shade tree allometries (optional)

Compute shade tree allometries, which are optional in the model. The function is available to easily change or add equations, not to be called directly.

Allometries(S, i)

Arguments

S
The list of class "Simulation" used by the DynACof model.

i
The index of the day.

Value

Any variable that is computed by the function. Default: DM_Stem_Tree, Height_Tree, CrownProj_Tree, CrownRad_Tree, Crown_H_Tree, Trunk_H_Tree, LA_Tree and LAD_Tree.

Details

This function is called from the Tree parameter functions, and then by the model. In-depth details are available in Vezy (2017)

References

Vezy, R., Simulation de pratiques de gestion alternatives pour l’adaptation des plantations pérennes aux changements globaux, in École doctorale science de l'environnement, spécialité physique de l'environnement. 2017, UNIVERSITÉ DE BORDEAUX: Bordeaux. p. 270.Link