Canopy and Soil boundary layer conductance (depreciated)

This function is not used in the model anymore, but is still available in the package for compatibility with older versions. Please use other functions from the package. This function assumes two aerodynamic conductances in series:

from the atmosphere to the canopy, based on Van de Griend (1989). This is actually two conductances, one in the inertial sublayer and one in the roughness sublayer.

within the canopy to the soil, based on Choudhury & Monteith (1988).

GBCANMS(WIND, ZHT, TREEH, Z0 = TREEH * 0.1, ZPD = TREEH * 0.75,
  GBCANMS1MIN = 0.0123, VONKARMAN = Constants()$vonkarman)

Arguments

WIND
Average daily windspeed (m s-1)

ZHT
Wind measurement height (m)

TREEH
Average tree height (m)

Z0
Roughness length (m), default to 0.1*TREEH

ZPD
Zero-plane displacement (m), defaults to 0.75*TREEH

GBCANMS1MIN
Minimum allowed atmosphere to canopy conductance (mol m-2 s-1), default to 0.0123

VONKARMAN
Von Karman constant, default to Constants()$vonkarman, 0.41.

Value

A list of three :

ustar

The friction velocity (m s-1)

Canopy

Atmosphere to canopy boundary layer conductance (mol m-2 s-1)

Soil

Canopy to soil boundary layer conductance (mol m-2 s-1)

Details

The defaults for Z0 and ZPD are computed very simply. Other simple formulations:

Lettau (1969) proposed an other way: $$Z0= 0.5 . h* . s / S$$ where h* is canopy height, s is the average silhouette area (projected area of the tree on a vertical plane) and S the specific area, with \(S= A/n\), where A is the total plot area and n the number of trees.

For ZPD, Verhoef et al. (1997) said \(d= 0.67\). h is a good proxy without any prior knowledge.

References

Van de Griend, A. A. and J. H. Van Boxel (1989). "Water and surface energy balance model with a multilayer canopy representation for remote sensing purposes." Water Resources Research 25(5): 949-971.

Choudhury, B. and J. Monteith (1988). "A four‐layer model for the heat budget of homogeneous land surfaces. " Quarterly Journal of the Royal Meteorological Society 114(480): 373-398.

See also

Examples

# Canop conductance of a forest:
GBCANMS(WIND = 3, ZHT = 26, TREEH = 25)$Canopy

#> [1] 6.208687

# And the conductance above its soil layer:
GBCANMS(WIND = 3, ZHT = 26, TREEH = 25)$Soil

#> [1] 0.05887955