Bulk aerodynamic conductance

Compute the aerodynamic conductance for sensible and latent heat above the canopy following Van de Griend and Van Boxel (1989).

G_bulk(Wind, ZHT, Z_top, Z0 = Z_top * 0.1, ZPD = Z_top * 0.75,
  alpha = 1.5, ZW = ZPD + alpha * (Z_top - ZPD), LAI, extwind = 0,
  vonkarman = Constants()$vonkarman)

Arguments

Wind
Average daily wind speed above canopy (m s-1)

ZHT
Wind measurement height (m)

Z_top
Average canopy height of the taller crop (m)

Z0
Roughness length (m). Default: 0.1*Z_top

ZPD
Zero-plane displacement (m), Default: 0.75*Z_top

alpha
Constant for diffusivity at top canopy. Default: 1.5 following Van de Griend et al (1989).

ZW
Top height of the roughness sublayer (m). Default: ZPD+alpha*(Z_top-ZPD)

LAI
Leaf area index of the upper layer (m2 leaf m-2 soil)

extwind
Extinction coefficient. Default: 0, no extinction.

vonkarman
Von Karman constant, default to Constants()$vonkarman, 0.41.

Value

G_bulk

The bulk aerodynamic conductance (m s-1)

Details

alpha can also be computed as: $$alpha=\frac{zw-d}{Z_{top}-d}$$ The bulk aerodynamic conductance \(ga_{bulk}\) is computed as follow: $$ga_{bulk}=\frac{1}{r1+r2+r3}$$ where r1, r2 and r3 are the aerodynamic resistances of the inertial sublayer, the roughness sublayer and the top layer of the canopy respectively. Because wind speed measurements are more often made directly in the roughness sublayer, the resistance in the inertial sublayer r1 is set to 0 though. r2 and r3 are computed using the equation 43 of Van de Griend and Van Boxel (refer to the pdf of the web version of the help file for Latex rendering) : $$r2=\int_{zh}^{zw}\frac{1}{K''}$$ with $$K''= kU_*(z_w-d)$$ And: $$r3=\int_{(z2+z1)/2}^{zh}\frac{1}{K'''}\mathrm{d}z$$ with $$K'''= U_z\frac{K_h}{U_h}$$

Integration of r2 and r3 equations give: $$\frac{(\ln(ZPD-ZW)^2-\ln(ZPD-Z2)^2)}{(2kU_*)}$$ simplified in: $$r2= \frac{1}{kU_*}\ln(\frac{ZPD-ZW}{ZPD-Z2})$$ and finaly: $$r3= \frac{Uh}{Kh}\ln(\frac{Uh}{U_{interlayer}})$$

References

Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model with a multilayer canopy representation for remote sensing purposes. Water Resources Research, 1989. 25(5): p. 949-971.

See also

Examples

# The bulk aerodynamic conductance for a coffee plantation managed in agroforestry system:
G_bulk(Wind=3,ZHT=25,Z_top=24,LAI = 0.5,extwind = 0.58)

#> [1] 1.046699