Canopy layer to canopy layer aerodynamic conductance

Compute the aerodynamic conductance for sensible and latent heat between canopy layers following Van de Griend and Van Boxel (1989).

G_interlay(Wind, ZHT, Z_top, Z0 = Z_top * 0.1, ZPD = Z_top * 0.75,
  alpha = 1.5, ZW = ZPD + alpha * (Z_top - ZPD), LAI_top, LAI_bot,
  extwind = 0, vonkarman = Constants()$vonkarman)

Arguments

Wind
Average daily wind speed above canopy (m s-1)

ZHT
Wind measurement height (m)

Z_top
Average canopy height of the taller crop (m)

Z0
Roughness length (m). Default: 0.1*Z_top

ZPD
Zero-plane displacement (m), Default: 0.75*Z_top

alpha
Constant for diffusivity at top canopy. Default: 1.5 following Van de Griend et al (1989).

ZW
Top height of the roughness sublayer (m). Default: ZPD+alpha*(Z_top-ZPD)

LAI_top
Leaf area index of the upper layer (m2 leaf m-2 soil).

LAI_bot
Leaf area index of the layer below the upper layer (m2 leaf m-2 soil).

extwind
Extinction coefficient. Default: 0, no extinction.

vonkarman
Von Karman constant, default to Constants()$vonkarman, 0.41.

Value

g_af

The aerodynamic conductance of the air between two canopy layers (m s-1)

Details

alpha can also be computed as: $$alpha=\frac{zw-d}{Z_{top}-d}$$ The aerodynamic conductance between canopy layers is computed as: $$g_{af}= \frac{1}{\frac{U_h}{K_h}\ln(U_{mid}/U_{inter})}$$ where usually \(U_{mid}\) is the wind speed at (median) cumulated LAI between the top and the soil, and \(U_{inter}\) the wind speed at the height between the two canopy layers. In this function, \(U_{mid}\) and \(U_{inter}\) are computed relative to the leaf area instead of the height of the vegetation layers.

References

Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model with a multilayer canopy representation for remote sensing purposes. Water Resources Research, 1989. 25(5): p. 949-971.

See also

Examples

# G_af for a coffee plantation managed in agroforestry system:
G_interlay(Wind = 3,ZHT = 25,Z_top = 2,LAI_top = 0.5,LAI_bot = 4)

#> [1] Inf