Canopy to soil aerodynamic conductance

Compute the aerodynamic conductance for sensible and latent heat between the center of the lowest canopy layer and the soil surface following Van de Griend and Van Boxel (1989).

G_soilcan(Wind, ZHT, Z_top, Z0 = Z_top * 0.1, ZPD = Z_top * 0.75,
  alpha = 1.5, ZW = ZPD + alpha * (Z_top - ZPD), LAI_top, LAI_bot,
  extwind = 0, vonkarman = Constants()$vonkarman)

Arguments

Wind
Average daily wind speed above canopy (m s-1)

ZHT
Wind measurement height (m)

Z_top
Average canopy height of the taller crop (m)

Z0
Roughness length (m). Default: 0.1*Z_top

ZPD
Zero-plane displacement (m), Default: 0.75*Z_top

alpha
Constant for diffusivity at top canopy. Default: 1.5 following Van de Griend et al (1989).

ZW
Top height of the roughness sublayer (m). Default: ZPD+alpha*(Z_top-ZPD)

LAI_top
Leaf area index of the upper layer (m2 leaf m-2 soil).

LAI_bot
Leaf area index of the layer below the upper layer (m2 leaf m-2 soil).

extwind
Extinction coefficient. Default: 0, no extinction.

vonkarman
Von Karman constant, default to Constants()$vonkarman, 0.41.

Value

\(g_a0\)

The aerodynamic conductance of the air between the lowest canopy layer and the soil surface (m s-1)

Details

alpha can also be computed as: $$alpha=\frac{zw-d}{Z_{top}-d}$$ The aerodynamic conductance between the lowest canopy layer and the soil is computed as: $$g_{a0}= \frac{1}{\frac{U_h}{K_h}\ln(U_{mid}/U_{0})}$$ where \(U_{mid}\) is the wind speed at median cumulated LAI between the top and the soil, and \(U_0\) the wind speed at soil surface.

References

Van de Griend, A.A. and J.H. Van Boxel, Water and surface energy balance model with a multilayer canopy representation for remote sensing purposes. Water Resources Research, 1989. 25(5): p. 949-971.

See also

Examples

# G_a0 for a coffee plantation managed in agroforestry system:
G_soilcan(Wind=1,ZHT=25,Z_top=24,LAI_top = 0.5,LAI_bot = 4,extwind=0.58)

#> [1] 1.172889