Leaf boundary layer conductance for heat

Compute the bulk leaf boundary layer conductance for heat using the wind speed, the leaf dimension, and leaf area distribution following Jones (1992) or Leuning et al. (1995).

Gb_h(Wind, wleaf = 0.068, LAI_lay, LAI_abv, extwind = 0, Z_top, ZHT,
  Z0 = Z_top * 0.1, ZPD = Z_top * 0.75, alpha = 1.5, ZW = ZPD + alpha *
  (Z_top - ZPD), Tleaf = NULL, Tair = NULL, Dheat = Constants()$Dheat,
  formulation = c("Jones_1992", "Leuning_1995"))

Arguments

Wind
Average daily wind speed above canopy (m s-1)

wleaf
Average leaf width (m)

LAI_lay
Leaf area index of the layer (m2 leaves m-2 soil)

LAI_abv
Cumulated leaf area index above the layer (m2 leaves m-2 soil)

extwind
Extinction coefficient. Default: 0, no extinction.

Z_top
Average canopy height of the taller crop (m)

ZHT
Wind measurement height (m)

Z0
Roughness length (m). Default: 0.1*Z_top

ZPD
Zero-plane displacement (m), Default: 0.75*Z_top

alpha
Constant for diffusivity at top canopy. Default: 1.5 following Van de Griend et al (1989).

ZW
Top height of the roughness sublayer (m). Default: ZPD+alpha*(Z_top-ZPD)

Tleaf
Leaf temperature (deg C). Only needed if formulation="Leuning_1995"

Tair
Canopy air temperature (deg C). Only needed if formulation="Leuning_1995"

Dheat
Molecular diffusivity for heat (m2 s-1). Default to Constants()$Dheat. Only needed if formulation="Leuning_1995"

formulation
The formulation used to compute \(Gb_h\)

Value

Gb

The leaf boundary layer conductance for heat (m s-1)

Details

The leaf boundary layer conductance for heat can be transformed into leaf boundary layer conductance for water vapour as follow: $$Gb_w= 1.075*gb_h$$ Note that \(Gb_w\) should be doubled for amphistomatous plants (stomata on both sides of the leaves).

References

Leuning, R., et al., Leaf nitrogen, photosynthesis, conductance and transpiration: scaling from leaves to canopies. Plant, Cell & Environment, 1995. 18(10): p. 1183-1200.

Mahat, V., D.G. Tarboton, and N.P. Molotch, Testing above‐ and below‐canopy represetations of turbulent fluxes in an energy balance snowmelt model. Water Resources Research, 2013. 49(2): p. 1107-1122.

See also

Examples

# Gb for a coffee plantation managed in agroforestry system:
Gb_h(Wind=3,wleaf=0.068,LAI_lay=4,LAI_abv=0.5,ZHT=25,Z_top=24,extwind=0.58)

#> [1] 0.03091239