Leaf boundary layer conductance for heat under forced and free convection

Compute the boundary layer conductance for heat - single sided, forced or free convection using the Leuning et al. (1995) equations.

Gb_hForced(Wind, wleaf = 0.068)

Gb_hFree(Tair, Tleaf, wleaf = 0.068, Dheat = Constants()$Dheat)

Arguments

Wind
Wind speed (m s-1)

wleaf
Leaf width (m). Default to 0.068

Tair
Average daily air temperature (deg C)

Tleaf
Average daily leaf temperature (deg C)

Dheat
Molecular diffusivity for heat (m2 s-1). Default to: Constants()$Dheat.

Value

\(Gb_hForced\)

Leaf boundary layer conductance for heat under forced convection (m s-1)

\(Gb_hFree\)

Leaf boundary layer conductance for heat under free convection (m s-1)

References

Leuning, R., et al., Leaf nitrogen, photosynthesis, conductance and transpiration: scaling from leaves to canopies. Plant, Cell & Environment, 1995. 18(10): p. 1183-1200.

See also

Examples

Gb_hForced(Wind=3)

#> [1] 0.01992633

Gb_hFree(Tair= 25,Tleaf= 26)

#> [1] 0.002367616