Import model parameters

Import the parameters from local files, or from default (from package data)

Import_Parameters(path = NULL, Names = list(Site = "Site.R", Soil =
  "Soil.R", Coffee = "Coffee.R", Tree = NULL))

Arguments

path
The path to the parameter files folder. If NULL, take the default files from the package

Names
List of the file names. Default: list(Site="Site.R",Soil="Soil.R",Coffee="Coffee.R",Tree=NULL)

Value

A list of all input parameters for DynACof

Details

For the full list of parameters and the format of the parameter files, see site. The function return the parameter files path in the list, to access it, see example.

See also

Examples

# Importing defaults:
Parameters= Import_Parameters()
# Parameters files path:
Parameters$files

#> $Site
#> [1] "Default from package function"
#> 
#> $Soil
#> [1] "Default from package function"
#> 
#> $Coffee
#> [1] "Default from package function"
#> 
#> $Tree
#> [1] "Default from package function"
#>