MAESPA metamodels

Several variables can be highly impacted by the 3D canopy structure. Hence, some variables are computed using metamodels from the MAESPA model (see details). These functions are example functions for metamodels and allometries. They are imported by the parameter files for a default simulation. Any simulation on different conditions than the one used as in the package default should have custom metamodels, and/or allometries. The user can write the customized functions directly in the parameter files were these functions are called. For example, to replace the light extinction coefficient computation, the user should write a customized function instead of the call to Light_extinction_K() on the line k= Light_extinction_K. These functions are helper functions available to the user to use as a template, but not to call them directly.

Light_extinction_K(S, i)

Metamodels_soil(S, i)

Arguments

S
The global list of class "Simulation" used by the DynACof model.

i
The day of interest.

Value

\(K_Dif_Tree\)

Shade tree diffuse light coefficient

\(K_Dir_Tree\)

Shade tree direct light coefficient

\(Rn_Soil\)

Soil net radiation (MJ m-2 d-1)

Details

The purpose of using metamodels in DynACof is to make this two-layer dynamic crop model able to consider the 3D canopy heterogeneity effect on fundamental processes, as if it was with a MAESPA-growth model coupling. The main process impacted by canopy complexity is the light absorbed by the plants (Charbonnier et al., 2013). Indeed, in dynamic crop models, the absorbed photosynthetically active radiation (APAR) by the canopy is often computed using the simple Beer-Lambert’s law or a derivative, with a variable leaf area index in time, but a constant extinction coefficient (Van Oijen et al., 2010b). However, heterogeneous canopies such as the shade trees in AFS coffee plantations tends to violate the assumption of a constant value for the diffuse (\(K_{dif_{Tree}}\)) and direct (\(K_{dir_{Tree}}\)) light extinction coefficients, because the spatial distribution of the leaf area is not uniform (high gap fraction) and because the leaf area density can change along time with foliage aggregation (Sampson and Smith, 1993;Sinoquet et al., 2007). Secondly, a comparison between coffee planted in monoculture and under agroforestry system showed that canopy complexity affected canopy temperature, water, and energy partitioning (Vezy et al., 2018), and probably photosynthesis because it is related to light interception and transpiration through stomatal conductance. Therefore, we derived metamodels from MAESPA for the diffuse (\(K_{dif_{Tree}}\)) and the direct (\(K_{dir_{Tree}}\)) shade tree light extinction coefficients, the light use efficiency (LUE, gC MJ-1), the coffee canopy temperature (Tcan, deg Celsius) and leaf water potential (MPa), the transpiration (T_x, mm) and plant sensible heat flux (H_x). The coffee layer was considered homogeneous enough to compute constant extinction coefficients derived from the MAESPA simulation, and the partitioning parameter between soil sensible and latent flux was also adjusted using MAESPA outputs. MAESPA is a 3D explicit model for energy, carbon and water fluxes simulation, for further details, see: Vezy et al. (2018), or the MAESPA website.

References

See Vezy (2017)

See also

Examples

# Creating a dummy list for use:
S= list(Sim= data.frame(K_Dif_Tree= rep(NA_real_,10),
        K_Dir_Tree= rep(NA_real_,10),LAD_Tree= rnorm(10,3,0.5)))
# Calling the function:
Light_extinction_K(S,1:10)