Evapotranspiration

Compute the daily evaporation or transpiration of the surface using the Penman-Monteith equation.

PENMON(Rn, Wind, Tair, ZHT, Z_top, Pressure, Gs, VPD, LAI, extwind = 0,
  wleaf = 0.068, Parameters = Constants())

Arguments

Rn
Net radiation (MJ m-2 d-1)

Wind
Wind speed (m s-1)

Tair
Air temperature (Celsius degree)

ZHT
Wind measurement height (m)

Z_top
Canopy top height (m)

Pressure
Atmospheric pressure (hPa)

Gs
Stomatal conductance (mol m-2 s-1)

VPD
Vapor pressure deficit (kPa)

LAI
Leaf area index of the upper layer (m2 leaf m-2 soil)

extwind
Extinction coefficient. Default: 0, no extinction.

wleaf
Average leaf width (m)

Parameters
Constant parameters, default to Constants, if different values are needed:

Cp

specific heat of air for constant pressure (J K-1 kg-1)

Rgas

universal gas constant (J mol-1 K-1)

Kelvin

conversion degree Celsius to Kelvin

H2OMW

conversion from kg to mol for H2O (kg mol-1)

Value

\(ET\), the daily (evapo|transpi)ration (mm d-1)

Details

The daily evapotranspiration is computed using the Penman-Monteith equation, and a set of conductances as : $$ET=\frac{\Delta\cdot Rn\cdot10^6+\rho\cdot Cp\cdot\frac{VPD}{10\ }\cdot GH}{\ \Delta+\frac{\gamma}{\lambda\ }\cdot(1+\frac{GH}{GV})}\ $$ where \(\Delta\) is the slope of the saturation vapor pressure curve (kPa K-1), \(\rho\) is the air density (kg m-3), \(GH\) the canopy boundary layer conductance (m s-1), \(\gamma\) the psychrometric constant (kPa K-1) and \(GV\) the boundary + stomatal conductance to water vapour (m s-1). To simulate evaporation, the input stomatal conductance \(Gs\) can be set to nearly infinite (e.g. \(Gs= 1\cdot e^9\)).

Note

If wind=0, it is replaced by a low value of 0.01

References

Allen R.G., Pereira L.S., Raes D., Smith M., 1998: Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56.

See also

Examples

# leaf evaporation of a forest :
PENMON(Rn= 12, Wind= 0.5, Tair= 16, ZHT= 26, Z_top= 25, Pressure= 900, Gs = 1E09, VPD= 2.41,
       LAI=3, extwind= 0.58, wleaf=0.068)

#> [1] 2.411908