Continuous percentage of living tissue distribution

Help compute the percentage of living tissue in organs according to age

Paliv_dis(Age_Max, P_Start, P_End, k)

Arguments

Age_Max
Maximum age of the organ (year)

P_Start
Percentage of living tissue at first age (% of dry mass)

P_End
Percentage of living tissue at last age (% of dry mass)

k
Rate between P_Start and P_End

Value

Living tissue at each age

(In % of organ dry mass)

Details

The percentage of living tissue is computed as follows: $$P_End+((P_Start-P_End)*exp(seq(0,-k,length.out = Age_Max)))$$

Examples

Paliv_dis(40,1,0.05,5)

#>    Age PaliveStem_Tree
#> 1    1           1.000
#> 2    2           0.886
#> 3    3           0.785
#> 4    4           0.697
#> 5    5           0.619
#> 6    6           0.550
#> 7    7           0.490
#> 8    8           0.437
#> 9    9           0.391
#> 10  10           0.350
#> 11  11           0.314
#> 12  12           0.282
#> 13  13           0.254
#> 14  14           0.229
#> 15  15           0.208
#> 16  16           0.189
#> 17  17           0.172
#> 18  18           0.157
#> 19  19           0.145
#> 20  20           0.133
#> 21  21           0.123
#> 22  22           0.114
#> 23  23           0.107
#> 24  24           0.100
#> 25  25           0.094
#> 26  26           0.089
#> 27  27           0.084
#> 28  28           0.080
#> 29  29           0.076
#> 30  30           0.073
#> 31  31           0.070
#> 32  32           0.068
#> 33  33           0.066
#> 34  34           0.064
#> 35  35           0.062
#> 36  36           0.061
#> 37  37           0.059
#> 38  38           0.058
#> 39  39           0.057
#> 40  40           0.056