Daily net radiation

Compute the daily net radiation of the system using incident radiation, air temperature, wind speed, relative humidity and the albedo. A clear description of this methodology can be found in Allen et al. (1998) or in An et al. (2017).

Rad_net(
  DOY,
  RAD,
  Tmax,
  Tmin,
  VPD,
  Rh = NULL,
  Latitude,
  Elevation,
  albedo,
  sigma = Constants()$sigma,
  Gsc = Constants()$Gsc
)

Arguments

DOY
Ordinal day, which is the day of year from 1st January (\(day\))

RAD
Incident daily total radiation (\(MJ\ m^{-2} d^{-1}\))

Tmax
Maximum daily air temperature (celsius degree)

Tmin
Minimum daily air temperature (celsius degree)

VPD
Mean daily Vapor Pressure Deficit (\(hPa\)), only needed if Rh is missing

Rh
Average daily relative humidity (%)

Latitude
Latitude (\(deg\))

Elevation
Elevation (\(m\))

albedo
Shortwave surface albedo (-)

sigma
Stefan-Boltzmann constant (\(W\ m^{-2} K^{-4}\)), default to Constants()$sigma.

Gsc
Solar constant (\(W\ m^{-2}\)), default to Constants()$Gsc (= 1367).

Value

\(Rn\), the daily net radiation (\(MJ\ m^{-2} d^{-1}\))

Details

The daily net radiation is computed using the surface albedo. This method is only a simple estimation. Several parameters (ac, bc, a1 and b1) are taken from Evett et al. (2011). The net radiation is computed as: $$Rn=(1-\alpha)\cdot RAD-(ac\cdot\frac{RAD}{Rso}+bc)\cdot(a1+b1\cdot ea^{0.5})\cdot\sigma\cdot\frac{T_{\max}^4+T_{\min}^4}{2}$$ And is derived from the equation : $$Rn= (1-\alpha)\cdot RAD-Rln$$ where \(Rln\) is the net upward longwave radiation flux, \(\alpha\) is the albedo, \(R_{so}\) the daily total clear sky solar irradiance, computed as follow: $$R_{so}= (0.75+0.00002\cdot Elevation)\cdot R{sa}$$ where \(R_{sa}\) is the daily extra-terrestrial radiation, computed using Rad_ext(). The actual vapor pressure \(ea\) can be computed using either VPD or the relative humidity and the maximum and minimum daily temperature. If both are provided, Rh will be used.

References

An, N., S. Hemmati, and Y.-J. Cui, Assessment of the methods for determining net radiation at different time-scales of meteorological variables. Journal of Rock Mechanics and Geotechnical Engineering, 2017. 9(2): p. 239-246.

Examples

# Daily net radiation on january 1st at latitude 9 N :
Rad_net(DOY= 1,RAD= 5,Tmax= 16,Tmin= 10,VPD=1.05,
        Latitude=9,Elevation=1000,albedo=0.146)

#> [1] 4.270029