Run DynACof.jl (Julia version of DynACof)

Make a simulation using the julia version of DynACof, DynACof.jl. This implementation is much faster, with run-times ~100x faster. Need Julia installed on the computer and in the path.

dynacof.jl(
  Period = NULL,
  WriteIt = FALSE,
  Inpath = NULL,
  output_f = ".RData",
  Outpath = Inpath,
  Simulation_Name = "DynACof",
  FileName = list(Site = "site.jl", Meteo = "meteorology.txt", Soil = "soil.jl", Coffee
    = "coffee.jl", Tree = NULL),
  ...
)

Arguments

Period
Period of time to be simulated, see details. Default: NULL

WriteIt
If TRUE, write the outputs to disk using write.results(), see details. Default: FALSE

Inpath
Path to the input parameter list folder, Default: NULL (take package values)

output_f
Output format. If output_f = ".RData", the output list will be saved as a unique .RData file. Any other value: write the output list in several .csv and .txt files. Default: .RData

Outpath
Path pointing to the folder were the results will be writen, Default: Outpath = Inpath

Simulation_Name
Character name of the simulation file name if WriteIt = T. Default: "DynACof"

FileName
A list of input file names :

Site

Site parameters file name, see details. Default: 'site.jl'

Meteo

Meteo parameters file name, see details. Default: 'meteorology.txt'

Soil

Soil parameters file name, see details. Default: 'soil.jl'

Coffee

Coffee parameters file name, see details. Default: 'coffee.jl'

Tree

Shade tree parameters file name, see details. Default: NULL

...
Further arguments to pass to write.results().

Value

Return invisibly a list containing three objects (Parameters, Meteo and Sim). See DynACof() for more details.

Note

If the user needs default values from the package for FileName names, put their values to "package". Careful: Always call dynacof.jl_setup() before runnning dynacof.jl. See example.

Examples

if (FALSE) {
# Setting up julia + DynACof.jl:
dynacof.jl_setup()

# Running a simulation:
# First, downloading an example meteorology file:
met_file= tempfile()
m= "https://raw.githubusercontent.com/VEZY/DynACof.jl_inputs/master/meteorology.txt"
download.file(m,met_file)

S= dynacof.jl(Inpath= dirname(met_file), FileName=list(Site="package",Meteo=basename(met_file),
                                                       Soil="package",Coffee="package",Tree=NULL))
unlink(met_file)
# Plotting the coffee LAI along the simulation:
plot(S$Meteo$Date,S$Sim$LAI)
}