Setup Julia call

This function initializes Julia and the DynACof.jl package. It is time-consuming for the first time of execution because of the precompilation.

dynacof.jl_setup(..., dev_path = NULL)

Arguments

...
Parameters are passed down to JuliaCall::julia_setup()

dev_path
the path to the dev folder location (e.g. "C:/Users/User/.julia/dev") if a dev version of DynACof.jl has to be used.

Details

This function requires Julia to be installed on your computer, and to be present in your path (i.e. callable from anywhere using a terminal). The function will test if the DynACof.jl package is installed in Julia, and if not, it will install it for you. This process can take some time, so be patient the first time you use it. If you encounter some issue, you can browe the issues from the JuliaCall package, or fill one in the DynACof repository.

Note

If you run into issues at this step, try to install the development version of JuliaCall: remotes::install_github("Non-Contradiction/JuliaCall"). If it does not work either, try to open Julia from the terminal, and run this command: using Pkg; Pkg.add(["RCall","Suppressor","DynACof"]) If you are using a new version of R, please use the argument rebuild= TRUE (see JuliaCall::julia_setup()).

Examples

if (FALSE) {
# Can be time-consuming the first time. Requires Julia + DynACof.jl
DynACof::dynacof.jl_setup()
}