Default (example) parameter functions

These functions return the default parameter values for DynACof. They are mainly used invisibly as defaults from Import_Parameters, but can still be called by the user for conveniance (but not needed for a model run). The parameters are divided into four functions: site(), soil(), coffee(), and Tree().

site()

soil()

coffee()

Tree()

Value

A list of parameters needed for a DynACof simulation.

Details

These input functions are stored into R-script that can be customized at will by the user to fit new conditions. In that case, the user have to write the resulting functions to separate files corresponding to the site, the soil, the coffee layer and the tree layer. Then DynACof read them using the Inpath and FileName parameters. File templates with default parameter values and an example use are available from this separate github repository

Note

The Tree parameter file is optional. All other parameter files should have the same structure as the one from the default parameter files. Other parameters can be added to the files though.

site

The default site is a stand from the Aquiares farm, located in Costa Rica. It is a Coffea arabica plantation in agroforestry management under Erythrina poeppigiana shade trees. The plot is visible at this address, and a full desciption is available here and here.

soil

The default soil parameters.

coffee

The default coffee parameters. It is essentially a high density plantation (6666 coffee plants per hectares) of Coffea arabica var. Caturra that are pruned every year to sustain three resprouts per stump in average.

Tree

The default shade tree parameters. The shade trees in Aquiares were Erythrina poeppigiana planted at high density (250 trees ha-1) and pruned to optimize light transmitted to the Coffea, and were depressed in 2000 to a low density of ~7.4 trees ha-1. From 2000, these trees made a relatively large crown with an average height of 26 m in 2018 on this site. NB: the Tree parameter file is optional, and not needed for monospecific coffee plantations.

See also