Shade Tree subroutine

Make all computations for shade trees (similar to coffee, but no fruits) for the ith day by modifying the S list in place.

Shade.Tree(S, i)

No_Shade(...)

Arguments

S
The simulation list of class "Simulation".

i
The index of the day since the first day of the simulation.

Value

Nothing, modify the list of simulation S in place. See DynACof() for more details.

Note

This function shouldn't be called by the user. It is made as a subroutine so it is easier for advanced users to modify the code. No_Shade() is used as an empty function that is called when there are no shade trees.

See also