Default (example) parameter functions

These input parameters for the model are imported as functions that are executed and return the list of parameters. It is strongly recommended for clarity to write the parameters into separate files corresponding to the site, the soil, the coffee layer and the tree layer. These functions are provided to define the default parameters for DynACof, and are used as example. They are mainly used invisibly as defaults from Import_Parameters, but can still be called by the user for conveniance (but not needed for a model run). The parameters are divided into four functions.

site()

soil()

coffee()

Tree()

Value

A list of parameters needed for a DynACof simulation. The full list (and example files) is available on the github repository in the data/ folder.

Note

The Tree parameter file is optional. All other parameter files should have the same structure as the one from the default parameter files. Other parameters can be added to the files though.

site

The default site is a stand from the Aquiares farm, located in Costa Rica. It is a Coffea arabica plantation in agroforestry management under Erythrina poeppigiana shade trees. The plot is visible at this address, and a full desciption is available here and here.

soil

The default soil parameters.

coffee

The default coffee parameters. It is essentially a high density plantation (6666 coffee plants per hectares) of Coffea arabica var. Caturra that are pruned every year to sustain three resprouts per cep in average.

Tree

The default shade tree parameters. The shade trees in Aquiares are Erythrina poeppigiana planted at a low density of ~7.4 trees ha-1. These trees make a relatively large crown and have an average height of 26 m in 2018 on this site. NB: the Tree parameter file is optional, and not needed for monospecific coffee plantations.

See also